+++ tests/nbconvert_html5.md
@@ -1,1 +1,1 @@
---8<-- "nbconvert_html5/templates/semantic-forms/README.md"
+--8<-- "nbconvert_a11y/templates/semantic-forms/README.md"
